% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_skiplogic.R
\name{apply_skiplogic}
\alias{apply_skiplogic}
\title{Apply Skip Logic to Survey Data}
\usage{
apply_skiplogic(path = NULL, x = NULL)
}
\arguments{
\item{path}{A character string specifying the path to the JSON file containing skip logic rules. Default is NULL.}

\item{x}{A data frame or tibble containing the survey data. Default is NULL.}
}
\value{
A modified data frame with skip logic applied.
}
\description{
This function applies skip logic to a survey data frame based on the skip logic rules defined in a JSON file. It modifies the survey data frame by setting values to "NOT APPLICABLE" where the skip logic conditions are met.
}
\examples{
\dontrun{
# Example usage:
modified_data <- apply_skiplogic(path = "path/to/your/skiplogic.json", x = your_survey_data)
}
}
