% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_forward.R
\name{carry_forward}
\alias{carry_forward}
\title{Carry Forward Selected Responses Into a New Variable}
\usage{
carry_forward(
  x,
  parent_var,
  child_var,
  new_var,
  orig_resp,
  keep_resp,
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{parent_var}{A character string naming the variable whose selected
responses should be carried forward.}

\item{child_var}{A character string naming the original variable that
corresponds to the new variable.}

\item{new_var}{A character string naming the new variable to create.}

\item{orig_resp}{A vector of responses in \code{parent_var} that should be carried
forward into \code{new_var}.}

\item{keep_resp}{A vector of responses in \code{child_var} that should be kept
and not overwritten.}

\item{verbose}{}
}
\value{
The input data frame with a new variable added.
}
\description{
Creates a new variable by carrying forward selected responses from a base
variable into a new variable, while preserving specified responses from an
original variable.
}
\examples{
#' \dontrun{
df <- data.frame(
  q1 = c("Yes", "No", "DK", "Yes", NA),
  q2 = c(NA, "Refused", "No", "Maybe", "Yes")
)

carry_forward(
  x = df,
  parent_var = "q1",
  child_var = "q2",
  new_var = "q2_cf",
  orig_resp = c("Yes", "No"),
  keep_resp = c("Refused", "Maybe")
)
}
}
