% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{label_data}
\usage{
label_data(x = NULL, codebook = NULL)
}
\arguments{
\item{x}{a \code{\link[=data.frame]{data.frame()}} object containing the raw survey data}

\item{codebook}{a \code{\link[=data.frame]{data.frame()}} object containing columns \code{options}, \code{values}, \code{grp}, and \code{orders}:
\itemize{
\item \code{options} : raw data values in the dataset
\item \code{values} : label values for each defined raw value in \code{options}
\item \code{grp} : the grouping value for the raw data and values. Often this will be at a question level but grouping may cut across multiple questions for common responses (i.e. 'yn' for yes and no questions)
\item \code{orders} : the order that the labels should appear as factor levels
}}
}
\value{
a labelled \code{\link[=data.frame]{data.frame()}} object where the data values are replaced with the \code{values} column in \code{codebook} based on the data value defined by \code{options}.
}
\description{
label_data
}
\examples{
# Codebook data frame
cb = data.frame(
  options = c(1, 2, 1, 2, 3),
  values = c("Yes", "No", "Urban", "Rural", "Other"),
  grp = c("yn", "yn", "geo", "geo", "geo"),
  orders = c(1, 2, 1, 2, 3)
)

# Raw, unlabeled dataset
data = data.frame(
  yn = c(1, 1, 2),
  geo = c(3, 1, 2)
)

# Labeled output
label_data(data, cb)

}
