% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var_app_server.R
\name{recode_var_app_server}
\alias{recode_var_app_server}
\title{Server for the recode_var Shiny app}
\usage{
recode_var_app_server(input, output, session, data, data_name)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments.}

\item{data}{A data frame or tibble passed into the app.}

\item{data_name}{A character string naming the data object used in emitted code.}
}
\value{
Called for side effects.
}
\description{
Server for the recode_var Shiny app
}
\keyword{internal}
