% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var_app_ui.R
\name{recode_var_app_ui}
\alias{recode_var_app_ui}
\title{UI for the recode_var Shiny app}
\usage{
recode_var_app_ui()
}
\value{
A Shiny UI object.
}
\description{
UI for the recode_var Shiny app
}
\keyword{internal}
