% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inr.R
\name{get_inr}
\alias{get_inr}
\title{get_inr}
\usage{
get_inr(
  mps = NULL,
  vars = NULL,
  include_skips = FALSE,
  skip_words = c("Skip", "SKIP", "skip", "REFUSED", "Refused", "refused")
)
}
\arguments{
\item{skip_words}{}
}
\description{
get_inr
}
