% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{recode_var}
\alias{recode_var}
\title{recode_var}
\usage{
recode_var(x, vars, new_var)
}
\arguments{
\item{x}{asdf}

\item{vars}{asdf}

\item{new_var}{asdf}
}
\description{
recode_var
}
\examples{
# hello
}
